fix(pwa,calendar): 更新no-op UpdateBannerを削除しTZ再解釈バグをlintで封じる
sw.jsがinstall時にskipWaitingするため「更新」ボタンは常にno-opだった
UpdateBannerを削除する（#2012）。壁時計DateをtoZonedTime/Intl.DateTimeFormat
で誤ってinstant再解釈する同型バグが2度発生したため、正しい変換helperを
lib/date/timezoneに集約し、date-fns-tzのtoZonedTime直接importをESLintで
禁止する（#2017）。

Closes #2012
Closes #2017
